requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
certifi==2025.11.12
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2025.6
Django==4.2.26
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
django-request-id==1.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
environ==1.0
flake8==7.3.0
gunicorn==23.0.0
inflection==0.5.1
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-memcached==1.62
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
sentry-sdk==2.44.0
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
